# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClipForge API - AI-powered YouTube to Shorts pipeline
 * OpenAPI spec version: 0.1.0
 */
import type { AccountPlatform } from "./accountPlatform";

export interface Account {
  id: number;
  platform: AccountPlatform;
  accountName: string;
  isConnected: boolean;
  createdAt: Date;
}