# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClipForge API - AI-powered YouTube to Shorts pipeline
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}